Fix turning on autosave w/o core reload
Previously, in order to enable autosave, you'd have to turn it on
using the OSD, then Save Settings, and then reload the core. This
commit fixes a bug preventing autosave from working as soon as it
was turned on. The problem was that a flag was not being reset when
one of the state machines exited if autosave was off, this
prevented the state machine from restarting properly the next time
the OSD was opened.